reduce servo range
The linear actuators used on the Newton robots deactivate if the servo
pulse length is shorter than 1 ms.

Previously the code placed the minimum at 1 ms, which, due to noise,
could cause the actuators to turn off and for steering to become loose.
To prevent this, a small buffer of 50 us has been added.

Although a buffer of 100 us would be preferable, due to the design of
the Servo library, the largest value valid for the minimum pulse width
is 1052 us.